increase allowed body limit and cap custom locations to 50k for now

+++ backend/routes/maps/updateCustomMap.ts
@@ -2,6 +2,8 @@ import { ObjectId } from 'mongodb'
 import { NextApiRequest, NextApiResponse } from 'next'
 import { calculateMapScoreFactor, collections, getUserId, throwError } from '@backend/utils'
 import { LocationType } from '@types'
+import { MAX_ALLOWED_CUSTOM_LOCATIONS } from '@utils/constants/random'
+import { formatLargeNumber } from '@utils/helpers'
 
 type ReqBody = {
   name?: string
@@ -72,6 +74,10 @@ const updateCustomMap = async (req: NextApiRequest, res: NextApiResponse) => {
 
   // HALP -> Really shouldn't be deleting locations and then inserting new
   if (locations) {
+    if (locations.length > MAX_ALLOWED_CUSTOM_LOCATIONS) {
+      return throwError(res, 400, `The max locations allowed is ${formatLargeNumber(MAX_ALLOWED_CUSTOM_LOCATIONS)}`)
+    }
+
     // Removes old locations
     const removeResult = await collections.userLocations?.deleteMany({ mapId: new ObjectId(mapId) })
 
@@ -97,16 +103,16 @@ const updateCustomMap = async (req: NextApiRequest, res: NextApiResponse) => {
       }
 
       // Update map's score factor (since locations have changed)
-      // const scoreFactor = calculateMapScoreFactor(locations)
+      const scoreFactor = calculateMapScoreFactor(locations)
 
-      // const updateMap = await collections.maps?.updateOne(
-      //   { _id: new ObjectId(mapId) },
-      //   { $set: { scoreFactor: scoreFactor } }
-      // )
+      const updateMap = await collections.maps?.updateOne(
+        { _id: new ObjectId(mapId) },
+        { $set: { scoreFactor: scoreFactor } }
+      )
 
-      // if (!updateMap) {
-      //   return throwError(res, 400, 'Failed to save new map score factor')
-      // }
+      if (!updateMap) {
+        return throwError(res, 400, 'Failed to save new map score factor')
+      }
     }
   }
